query in db for name of dish or ingredients

+++ src/controllers/dishesControllers.js
@@ -1,4 +1,5 @@
 const knex = require('../database/knex')
+const sqlConnection = require('../database/sql')
 const AppError = require('../utils/AppError')
 const DiskStorage = require('../providers/diskStorage')
 
@@ -34,19 +35,19 @@ class DishesControllers{
         const dish_id = request.params.id
         const {name, description, category, price, ingredients} = request.body
         const dish = await knex('dishes').where({id: dish_id}).first()
-  
+
         if(!dish) throw new AppError('Prato não existente')
 
         dish.name = name ?? dish.name
         dish.description = description ?? dish.description
         dish.category = category ?? dish.category
         dish.price = price ?? dish.price
-        
-        if (ingredients?.length !== 0) {
+
+        if (ingredients?.length !== 0 && ingredients !== undefined) {
     
             await knex('ingredients').where({dish_id}).delete()
         
-            const ingredientsUpdate = ingredients.map(name => {
+            const ingredientsUpdate = ingredients?.map(name => {
                 return {
                     name,
                     dish_id,
@@ -69,15 +70,33 @@ class DishesControllers{
     }
     
     async Show(request,response){
-        const dishes = await knex('dishes')
+        const {id} = request.params
+        const dishes = await knex('dishes').where({id})
+        
         return response.json(dishes)
     }
     
-    async Index(request,response){
-        const dish_id = request.params.id
-        const dishes = await knex('dishes').where({id: dish_id}).first()
+    async Index(request, response) {
+        const { name } = request.query;
+        let dishes;
+      
+        dishes = await knex('dishes')
+          .select(['*'])
+          .where('dishes.name', 'like', `%${name}%`)
+          .orderBy('dishes.name');
+      
+        if (dishes.length === 0) {
+
+          dishes = await knex('dishes')
+            .select(['dishes.*'])
+            .distinct()
+            .innerJoin('ingredients', 'dishes.id', '=', 'ingredients.dish_id')
+            .where('ingredients.name', 'like', `%${name}%`)
+            .orderBy('dishes.name');
+        }      
 
         return response.json(dishes)
+    
     }
 }
